requirements.txt
fastapi
uvicorn
sqlalchemy
langchain
openai
python-dotenv
python-jose
pydantic
passlib[bcrypt]
pymysql